qrスキャン機能を設定画面に移設
