+++ app/(tabs)/(stacks)/PrisonAreaScreen.tsx
@@ -23,9 +23,7 @@ function PrisonAreaScreen({ _userStore, _tagGameStore }: Props) {
   const isGameStartDone = useRef(false);
 
   const isGameMaster = () => {
-    return userStore
-      .getCurrentUser()
-      .isCurrentGameMaster(tagGameStore.getTagGame());
+    return userStore.isCurrentUserGameMaster(tagGameStore.getTagGame());
   };
 
   const storeGameStartSetting = async (gameId: string) => {
