+++ app/(tabs)/(stacks)/index.tsx
@@ -1,19 +1,29 @@
 import { inject, observer } from "mobx-react";
 import { router } from "expo-router";
-import { View } from "react-native";
+import { Text, View } from "react-native";
+import ReactNativeModal from "react-native-modal";
 
 import TagGameStore from "@/stores/TagGameStore";
 import UserStore from "@/stores/UserStore";
-import { putTagGames } from "@/utils/APIs";
+import { joinUser, patchDevices, putTagGames } from "@/utils/APIs";
 import { Button } from "@rneui/themed";
+import { useRef, useState } from "react";
+import { CameraView } from "expo-camera";
+import { IconSymbol } from "@/components/ui/IconSymbol";
+import TagGameModel from "@/models/TagGameModel";
 
 interface Props {
   _userStore?: UserStore;
   _tagGameStore?: TagGameStore;
 }
 
-function SettingScreen({ _tagGameStore }: Props) {
+function SettingScreen({ _userStore, _tagGameStore }: Props) {
+  const userStore = _userStore!;
   const tagGameStore = _tagGameStore!;
+
+  const [cameraVisible, setCameraVisible] = useState(false);
+
+  const firstScan = useRef(true);
 
   const gameStart = () => {
     tagGameStore.getTagGame().setIsGameStarted(true);
@@ -25,6 +35,38 @@ function SettingScreen({ _tagGameStore }: Props) {
     tagGameStore.getTagGame().setIsGameStarted(false);
     const tagGame = tagGameStore.getTagGame().toObject();
     putTagGames(tagGame);
+  };
+
+  // ゲームマスター以外の人(子)がゲーム情報をstoreにセットするための処理
+  const setDataSettings = async ({ data: gameId }: { data: string }) => {
+    // NOTE: カメラモーダルを閉じた際にtrueに戻します。
+    // NOTE: QRが画面上にある限り廉造スキャンしてしまうので最初のスキャン以外は早期リターンしている
+    if (!firstScan.current || !userStore.getCurrentUser().getDeviceId()) return;
+
+    firstScan.current = false;
+    console.log("ScanData: ", gameId);
+    setCameraVisible(false);
+    tagGameStore.getTagGame().setId(gameId);
+    await patchDevices(gameId, userStore.getCurrentUser().getDeviceId());
+
+    const updatedLiveUsers = await joinUser(
+      gameId,
+      userStore.getCurrentUser().getDeviceId(),
+    );
+
+    const tagGame = new TagGameModel({
+      id: gameId,
+      validAreas: tagGameStore.getTagGame().getValidAreas(),
+      liveUsers: updatedLiveUsers.liveUsers,
+      rejectUsers: [],
+      // TODO: ゲームマスターを取得できるようにしたい。現状は自分がゲームマスターでないことしかわからない
+      gameMasterDeviceId: "",
+      prisonArea: [],
+      policeUsers: [],
+      gameTimeLimit: "",
+      isGameStarted: null,
+    });
+    tagGameStore.putTagGame(tagGame);
   };
 
   const canGameStart = () => {
@@ -41,63 +83,129 @@ function SettingScreen({ _tagGameStore }: Props) {
       style={{ height: "100%", alignItems: "center", backgroundColor: "white" }}
     >
       <View style={{ gap: 100, height: "80%" }}>
-        <Button
-          color={
-            tagGameStore.getTagGame().getIsSetValidAreaDone()
-              ? "success"
-              : "error"
-          }
-          title="ゲーム範囲設定"
-          onPress={() => {
-            router.push("/ValidAreaScreen");
-          }}
-        ></Button>
-        <Button
-          color={
-            tagGameStore.getTagGame().getIsSetPrisonAreaDone()
-              ? "success"
-              : "error"
-          }
-          title="監獄エリア設定"
-          onPress={() => {
-            router.push("/PrisonAreaScreen");
-          }}
-        ></Button>
-        <Button
-          color={tagGameStore.getIsEditTeams() ? "success" : "error"}
-          title="チーム設定"
-          onPress={() => {
-            router.push("/TeamEditScreen");
-          }}
-        ></Button>
-        <Button
-          color={
-            tagGameStore.getTagGame().getGameTimeLimit() ? "success" : "error"
-          }
-          title="タイムリミット設定"
-          onPress={() => {
-            router.push("/GameTimeScreen");
-          }}
-        ></Button>
+        {userStore.isCurrentUserGameMaster(tagGameStore.getTagGame()) ? (
+          <>
+            <Button
+              color={
+                tagGameStore.getTagGame().getIsSetValidAreaDone()
+                  ? "success"
+                  : "error"
+              }
+              title="ゲーム範囲設定"
+              onPress={() => {
+                router.push("/ValidAreaScreen");
+              }}
+            ></Button>
+            <Button
+              color={
+                tagGameStore.getTagGame().getIsSetPrisonAreaDone()
+                  ? "success"
+                  : "error"
+              }
+              title="監獄エリア設定"
+              onPress={() => {
+                router.push("/PrisonAreaScreen");
+              }}
+            ></Button>
+            <Button
+              color={tagGameStore.getIsEditTeams() ? "success" : "error"}
+              title="チーム設定"
+              onPress={() => {
+                router.push("/TeamEditScreen");
+              }}
+            ></Button>
+            <Button
+              color={
+                tagGameStore.getTagGame().getGameTimeLimit()
+                  ? "success"
+                  : "error"
+              }
+              title="タイムリミット設定"
+              onPress={() => {
+                router.push("/GameTimeScreen");
+              }}
+            ></Button>
+            {/* </View> */}
+            {tagGameStore.getTagGame().getIsGameStarted() === true ? (
+              <Button
+                title={"ゲーム中止"}
+                color={"error"}
+                onPress={() => {
+                  gameCancel();
+                }}
+              />
+            ) : (
+              <Button
+                title={"ゲームスタート"}
+                color={"primary"}
+                disabled={!canGameStart()}
+                onPress={() => {
+                  gameStart();
+                }}
+              />
+            )}
+          </>
+        ) : (
+          <>
+            {/* <View style={{ gap: 100, height: "80%" }}> */}
+            <Button
+              color={
+                tagGameStore.getTagGame().getIsSetValidAreaDone()
+                  ? "success"
+                  : "error"
+              }
+              title="ゲーム範囲設定"
+              onPress={() => {
+                router.push("/ValidAreaScreen");
+              }}
+            />
+            <Button
+              type="solid"
+              onPress={() => {
+                setCameraVisible(true);
+              }}
+              title="ゲームに参加"
+              icon={<IconSymbol size={28} name={"camera"} color={"white"} />}
+            />
+            {/* </View> */}
+            <ReactNativeModal
+              style={{ margin: "auto" }}
+              isVisible={cameraVisible}
+            >
+              <View
+                style={{ backgroundColor: "white", width: 330, padding: 20 }}
+              >
+                <Text style={{ fontSize: 18 }}>
+                  {"QRを読み込ませてもらって\nゲームグループに参加しましょう!!"}
+                </Text>
+                <CameraView
+                  style={{
+                    width: 250,
+                    height: 300,
+                    marginHorizontal: "auto",
+                    marginBottom: 20,
+                  }}
+                  barcodeScannerSettings={{
+                    barcodeTypes: ["qr"],
+                  }}
+                  onBarcodeScanned={setDataSettings}
+                  facing={"back"}
+                />
+                <Button
+                  type="solid"
+                  color={"red"}
+                  onPress={() => {
+                    setCameraVisible(false);
+                    firstScan.current = true;
+                  }}
+                >
+                  閉じる
+                </Button>
+              </View>
+            </ReactNativeModal>
+          </>
+        )}
       </View>
-      {tagGameStore.getTagGame().getIsGameStarted() === true ? (
-        <Button
-          title={"ゲーム中止"}
-          color={"error"}
-          onPress={() => {
-            gameCancel();
-          }}
-        />
-      ) : (
-        <Button
-          title={"ゲームスタート"}
-          color={"primary"}
-          disabled={!canGameStart()}
-          onPress={() => {
-            gameStart();
-          }}
-        />
-      )}
     </View>
   );
 }
